+++ optbinning/scorecard/counterfactual/counterfactual_information.py
@@ -5,9 +5,9 @@ Counterfactual information.
 # Guillermo Navas-Palencia <g.navas.palencia@gmail.com>
 # Copyright (C) 2021
 
-from ...binning.binning_information import print_header
-from ...binning.binning_information import print_optional_parameters
-from ...binning.binning_information import print_solver_statistics
+from ...information import print_header
+from ...information import print_optional_parameters
+from ...information import print_solver_statistics
 from ...options import counterfactual_default_options
